Split out interfaces.
