+++ src/Nito.Disposables/IReferenceCountedDisposable.cs
@@ -1,0 +1,26 @@
+﻿using System;
+
+namespace Nito.Disposables
+{
+    /// <summary>
+    /// An instance that represents a reference count.
+    /// </summary>
+    public interface IReferenceCountedDisposable<out T> : IDisposable
+        where T : class, IDisposable
+    {
+        /// <summary>
+        /// Adds a weak reference to this reference counted disposable. Throws <see cref="ObjectDisposedException"/> if this instance is disposed.
+        /// </summary>
+        IWeakReferenceCountedDisposable<T> AddWeakReference();
+
+        /// <summary>
+        /// Returns a new reference to this reference counted disposable, incrementing the reference counter. Throws <see cref="ObjectDisposedException"/> if this instance is disposed.
+        /// </summary>
+        IReferenceCountedDisposable<T> AddReference();
+
+        /// <summary>
+        /// Gets the target object. Throws <see cref="ObjectDisposedException"/> if this instance is disposed.
+        /// </summary>
+        T Target { get; }
+    }
+}
